+++ src/panels/lovelace/editor/hui-badge-preview.ts
